src/third_party/blink/web_tests/external/wpt/workers/interfaces/SharedWorkerGlobalScope/onconnect-expected.txt

This is a testharness.js-based test.
FAIL onconnect assert_array_equals: property 1, expected "null" but got "[object Object]"
Harness: the test ran to completion.
